refactor: move factory into isolated module
