+++ code/ui.datagrid/test/SAMPLE/index.ts
@@ -1,3 +1,4 @@
 export * from './DATA';
+export * from './factory';
 export * from './getFunc';
 export * from './util';
